requirements.txt
Django>=4.2.0
psycopg2-binary>=2.9.0
qrcode[pil]>=7.4.0
Pillow>=10.0.0
python-decouple>=3.8
openpyxl>=3.1.0

